# src/TripBundle/Resources/api/openapi.yaml
openapi: "3.0.5"
info:
    version: 1.0.0
    title: Trips
    license:
        name: GNU AGPLv3
servers:
    -   url: /
security:
    -   BasicAuth: [ ]
paths:
...
    /trip/countries/:
        get:
            summary: Get list of countries
            operationId: getCountries
            tags:
                - countries
            responses:
                '200':
                    description: ok
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Countries"
                default:
                    description: unexpected error
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/DefaultResponse"
components:
    securitySchemes:
        BasicAuth:
            type: http
            scheme: basic
    schemas:
...
        Country:
            type: object
            properties:
                name:
                    type: string
                code:
                    type: string
                region:
                    type: string
        Countries:
            type: array
            items:
                $ref: "#/components/schemas/Country"
        DefaultResponse:
            type: object
            required:
                - code
                - message
            properties:
                code:
                    type: integer
                    format: int32
                message:
                    type: string